pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry]
name = "terok-executor"
version = "0.0.70"
description = "Single-agent task runner for hardened Podman containers"
readme = "README.md"
license = "Apache-2.0"
authors = ["terok maintainers"]
keywords = ["podman", "containers", "agent", "ai", "coding-agent"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Environment :: Console",
  "Operating System :: POSIX :: Linux",
  "Topic :: Software Development :: Quality Assurance",
]

packages = [
  { include = "terok_executor", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
terok-sandbox = {url = "https://github.com/terok-ai/terok-sandbox/releases/download/v0.0.62/terok_sandbox-0.0.62-py3-none-any.whl"}
"ruamel.yaml" = ">=0.18"
Jinja2 = ">=3.1"
tomli-w = ">=1.0"

[tool.poetry.group.dev.dependencies]
pre-commit = ">=4.0"
ruff = ">=0.8"
bandit = ">=1.7"
docstr-coverage = "^2.3.2"
complexipy = ">=5.0"
vulture = ">=2.12"
reuse = "^6.2.0"
tach = "^0.34.1"

[tool.poetry.group.test.dependencies]
pytest = ">=7.0"
pytest-cov = ">=4.0"
pytest-timeout = ">=2.0"

[tool.poetry.group.docs.dependencies]
mkdocs = ">=1.5,<2"
mkdocs-material = ">=9.0"
mkdocstrings = {extras = ["python"], version = ">=0.24"}
mkdocs-literate-nav = ">=0.6"
mkdocs-section-index = ">=0.3"
mkdocs-terok = {url = "https://github.com/terok-ai/mkdocs-terok/releases/download/v0.4.1/mkdocs_terok-0.4.1-py3-none-any.whl"}

[tool.poetry.scripts]
terok-executor = "terok_executor.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*", "*Tests"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
  "needs_podman: requires podman on the host",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",      # pycodestyle errors
  "W",      # pycodestyle warnings
  "F",      # pyflakes
  "I",      # isort
  "B",      # flake8-bugbear
  "C4",     # flake8-comprehensions
  "UP",     # pyupgrade
  "SIM",    # flake8-simplify
]
ignore = [
  "E501",   # line too long (handled by formatter)
  "B904",   # raise from in except
  "SIM102", # nested if statements
  "SIM105", # contextlib.suppress vs try/except/pass
  "SIM108", # use ternary operator
  "SIM117", # nested with statements
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["terok_executor"]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
pattern = "^v(?P<base>\\d+\\.\\d+\\.\\d+)$"
style = "pep440"

[tool.poetry-dynamic-versioning.substitution]
folders = [{path = "src"}]
